# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "openrouter · servicio",
    "version": "0.8.0",
    "summary": "Pasarela propia sobre OpenRouter y el router de Hugging Face, pensada para ser llamada por otras aplicaciones.",
    "description": "Esto no es una aplicación con interfaz: es un conjunto de servicios HTTP. La clave de OpenRouter vive solo en el backend y nunca sale de él. Cada aplicación consumidora tiene su propia clave, que se da de alta y se revoca sin tocar el despliegue.\n\nEl contrato sigue el de la API de OpenAI donde existe equivalente, así que cualquier cliente de OpenAI funciona cambiando la URL base. Lo propio del servicio va en rutas propias y en cabeceras `X-...`. Una ruta publicada no se rompe: lo incompatible iría a `/v2/`.\n\nLa versión de este documento acompaña a la etapa entregada: `0.6.0` es la etapa 6; `0.6.2` amplía `/v1/models` con descripción, precios completos, modalidades de salida y parámetros, sin cambiar nada de lo publicado; `0.6.4` entrega la etapa 9 (catálogo completo de OpenRouter y compra en el router de Hugging Face con el prefijo `hf:`), también sin cambiar nada: solo añade campos, parámetros y códigos. Las etapas 7 y 8 siguen pendientes.",
    "license": {
      "name": "Uso interno"
    }
  },
  "servers": [
    {
      "url": "https://apisor.oracle402.com",
      "description": "Producción (VPS)"
    },
    {
      "url": "https://openrouter-npiobject-labs.fly.dev",
      "description": "Previsualización de main (Fly.io)"
...
  "security": [
    {
      "claveDeServicio": []
    }
  ],
...
  "paths": {
...
    "/v1/estado": {
      "get": {
        "tags": [
          "Servicio"
        ],
        "summary": "Estado del servicio y de la clave",
        "description": "Comprueba de verdad que el backend habla con OpenRouter, y devuelve la etiqueta de la clave, su uso y su límite tal como los da el proveedor. No gasta crédito, así que sirve para vigilancia periódica.",
        "responses": {
          "200": {
            "description": "Conexión confirmada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estado"
                },
                "example": {
                  "ok": true,
                  "build": "87e6a89",
                  "modelo_defecto": "openai/gpt-oss-120b",
                  "modelo_multimodal": "google/gemini-3.1-flash-lite",
                  "clave": {
                    "label": "openrouter",
                    "usage": 0.0123,
                    "limit": 5,
                    "is_free_tier": false
                  },
                  "almacen": "sqlite",
                  "registros": 128,
                  "identidad": {
                    "nombre": "presupuestos",
                    "app_id": "app_27b1d11e45",
                    "administracion": false
                  },
                  "upstreams": {
                    "openrouter": true,
                    "hf": true
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "503": {
            "$ref": "#/components/responses/SinConfigurar"
          }
        }
      }
    },
...
    "/v1/videos/models": {
      "get": {
        "tags": [
          "Vídeo"
        ],
        "summary": "Catálogo de vídeo",
        "operationId": "listarModelosDeVideo",
        "description": "Los modelos de vídeo de OpenRouter y de Higgsfield en un solo formato. Los de Higgsfield solo salen si el despliegue tiene `HIGGSFIELD_API_KEY`. No gasta crédito. Caché de una hora por proveedor.",
        "parameters": [
          {
            "name": "upstream",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "openrouter",
                "higgsfield",
                "todos"
              ],
              "default": "todos"
            }
          },
          {
            "name": "tarea",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "texto_a_video",
                "imagen_a_video",
                "video_a_video",
                "alargar",
                "ampliar",
                "avatar"
              ]
            }
          },
          {
            "name": "texto",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max_precio_segundo",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "example": 0.1
          }
        ],
        "responses": {
          "200": {
            "description": "`upstreams` dice cuántos modelos trae cada proveedor (`null` si no tiene clave).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "object": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ModeloVideo"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "upstreams": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PeticionInvalida"
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "502": {
            "$ref": "#/components/responses/FalloDeProveedor"
          },
          "503": {
            "$ref": "#/components/responses/SinConfigurar"
          }
        }
      }
    },
    "/v1/videos": {
      "post": {
        "tags": [
          "Vídeo"
        ],
        "summary": "Generar o editar un vídeo",
        "operationId": "crearVideo",
        "description": "Encola un vídeo y responde `202` al momento con el trabajo; se consulta después en `GET /v1/videos/{id}`. El cuerpo es el del proveedor más `model`: en OpenRouter, `prompt`, `duration`, `resolution`, `aspect_ratio`, `generate_audio`, `frame_images`, `input_references`...; en Higgsfield, el del modelo elegido (`prompt`, `image_url`, `video_url`...).\n\nAntes de encolar se estima el coste y, con clave de aplicación, se comprueba que cabe en lo que le queda de presupuesto: si no, `402 presupuesto_insuficiente` sin gastar nada. Lo estimado cuenta en el presupuesto hasta que el vídeo termina; entonces se sustituye por el coste real (o 0 si falló). Pasa también por el cortacircuitos de bucles y la cuota por minuto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "model"
                ],
                "properties": {
                  "model": {
                    "type": "string"
                  }
                },
                "additionalProperties": true
              },
              "example": {
                "model": "google/veo-3.1-fast",
                "prompt": "Un perro corriendo por la playa al atardecer",
                "duration": 4,
                "resolution": "720p",
                "aspect_ratio": "16:9",
                "generate_audio": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Encolado.",
            "headers": {
              "X-Uso-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Presupuesto": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajoVideo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PeticionInvalida"
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "402": {
            "$ref": "#/components/responses/SinSaldo"
          },
          "429": {
            "$ref": "#/components/responses/DemasiadasPeticiones"
          },
          "502": {
            "$ref": "#/components/responses/FalloDeProveedor"
          },
          "503": {
            "$ref": "#/components/responses/SinConfigurar"
          }
        }
      }
    },
    "/v1/videos/{id}": {
      "get": {
        "tags": [
          "Vídeo"
        ],
        "summary": "Estado de un vídeo",
        "operationId": "consultarVideo",
        "description": "El trabajo, con su estado. Si no ha terminado se pregunta al proveedor (como mucho cada 3 segundos). Al terminar trae `urls` y el coste definitivo. Un vídeo de otra aplicación responde `404`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El trabajo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajoVideo"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "404": {
            "$ref": "#/components/responses/NoEncontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "claveDeServicio": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clave en `Authorization: Bearer <clave>`.\n\nHay dos tipos. La **clave de administración** es la del despliegue y puede con todo, incluidas las rutas de `/v1/apps`. Las **claves de aplicación** se dan de alta en `POST /v1/apps`, se revocan una a una y solo ven su propio gasto.\n\nNunca la pongas en código de navegador: llama desde el backend de tu aplicación. Las cuotas y presupuestos por aplicación llegan en la etapa 6."
      }
    },
...
    "responses": {
      "NoAutorizado": {
        "description": "Falta la clave de servicio o no es válida.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "ok": false,
              "error": {
                "message": "Falta la cabecera Authorization: Bearer <clave de servicio>.",
                "code": "sin_clave",
                "type": "authentication_error"
              }
            }
          }
        }
      },
      "PeticionInvalida": {
        "description": "El cuerpo no es un objeto JSON, o pide streaming, que aún no existe.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "ok": false,
              "error": {
                "message": "El streaming llega en la etapa 7; manda la consulta sin \"stream\".",
                "code": "streaming_no_disponible",
                "type": "invalid_request_error"
              }
            }
          }
        }
      },
...
      "SinConfigurar": {
        "description": "Al servicio le falta un secreto y no puede atender: `sin_configurar` si falta la clave de servicio o la de OpenRouter; `upstream_sin_configurar` si solo falta la de otro upstream (`HF_TOKEN`) y se pidió un modelo suyo.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "ok": false,
              "error": {
                "message": "El servicio no tiene clave de OpenRouter configurada; revisa los secretos de Fly.",
                "code": "sin_configurar",
                "type": "api_error"
              }
            }
          }
        }
      },
...
    }
  },
...
}